Feature: Add modal for CreateTransaction widget. Remove unused code & minor refactoring
