+++ lib/widgets/transactions_list.dart
@@ -6,7 +6,7 @@ import '../widgets/transaction_card.dart';
 class TransactionList extends StatelessWidget {
   final List<Transaction> transactions;
 
-  const TransactionList({required this.transactions, super.key});
+  const TransactionList(this.transactions, {super.key});
 
   @override
   Widget build(BuildContext context) {
